% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive_mixed_model.R
\name{additive_mixed_model}
\alias{additive_mixed_model}
\title{Additive Mixed Model}
\usage{
additive_mixed_model(model1, model2, model1_weight)
}
\arguments{
\item{model1}{Vector.}

\item{model2}{Vector.}

\item{model1_weight}{numeric. Weight of model1}
}
\description{
This function will return the weighted sum between two models of the same length
}
