% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{append_predicted_values}
\alias{append_predicted_values}
\title{Append Predicted Values}
\usage{
append_predicted_values(
  data,
  model,
  left_trunc = 1,
  right_trunc = left_trunc + length(data) - 1
)
}
\description{
This function will take in a left truncated data vector and append missing values according to the given model
}
