% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{calculate_label_coords}
\alias{calculate_label_coords}
\title{Label Coordinate Calculate Helper Function}
\usage{
calculate_label_coords(
  x_lower_bound,
  x_upper_bound,
  y_lower_bound,
  y_upper_bound,
  x_buffer = 0.5,
  y_buffer = 0.5,
  log_scale_x = FALSE,
  log_scale_y = FALSE
)
}
\arguments{
\item{x_lower_bound}{Numeric lowest value of x axis}

\item{x_upper_bound}{Numeric highest value of x axis}

\item{y_lower_bound}{Numeric lowest value of y axis}

\item{y_upper_bound}{Numeric highest value of y axis}

\item{x_buffer}{Numeric indicating location on x axis (0 - 1)}

\item{y_buffer}{Numeric indicating location on y axis (0 - 1)}

\item{log_scale_x}{Boolean indicating if x axis is log scale}

\item{log_scale_y}{Boolean indicating if y axis is log scale}
}
\description{
This function calculates coordinates for a plot given x and y bounds and location represented as percentage of plot area
}
