% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_to_sample.R
\name{data_table_to_sample}
\alias{data_table_to_sample}
\title{Data Table to Sample}
\usage{
data_table_to_sample(dt)
}
\arguments{
\item{dt}{DataFrame. Frequency dataframe}
}
\description{
Helper function which converts a frequency dataframe to a sample vector
}
