% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Distribution Function}
\usage{
Exponential(n, a, b)
}
\arguments{
\item{n}{Length of vector to be generated.}

\item{a}{Parameter of the Exponential distribution function}

\item{b}{Parameter of the Exponential distribution function}
}
\description{
This function generates a vector of n length of the Exponential distribution with parameters a and b.
}
\examples{
Exponential(100, 10000, 0.8)
}
