% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{fit_generalized_kw}
\alias{fit_generalized_kw}
\title{Fit Generalized KW}
\usage{
fit_generalized_kw(
  param_bounds,
  m,
  data,
  weighted_rt = FALSE,
  pmf_weight = 0,
  par_chunk = 100,
  par_chunk_size = 10,
  n_cores = 1,
  clust,
  left_trunc = 1,
  right_trunc = left_trunc + length(data) - 1
)
}
\description{
This function will take in a set of parameters and calculate goodness of fit based on the Psi criterion metric
}
