% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{fit_mixed_kw}
\alias{fit_mixed_kw}
\title{Fit Mixed Kolmogorov-Waring}
\usage{
fit_mixed_kw(
  param_bounds,
  d,
  data,
  weighted_rt = FALSE,
  pmf_weight = 0,
  par_chunk = 100,
  par_chunk_size = 10,
  n_cores = 12,
  clust,
  left_trunc = 1,
  right_trunc = left_trunc + length(data) - 1
)
}
\description{
Fit Mixed Kolmogorov-Waring
}
