% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Generalized_Pareto}
\alias{Generalized_Pareto}
\title{Generalized Pareto Distribution Function}
\usage{
Generalized_Pareto(k, theta, c, b, rho)
}
\arguments{
\item{k}{Length of vector to be generated}

\item{theta}{Parameter of the Generalized Pareto function}

\item{c}{Parameter of the Generalized Pareto function}

\item{b}{Parameter of the Generalized Pareto function}

\item{rho}{Parameter of the Generalized Pareto function}
}
\description{
Returns vector of length k of Generalized Pareto given a parameters theta, c, b , and rho
}
