% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Generalized_Pareto_calc_P0_delta}
\alias{Generalized_Pareto_calc_P0_delta}
\title{Generalized Pareto Distribution Function P0 with defined delta}
\usage{
Generalized_Pareto_calc_P0_delta(theta, c, b, rho, delta)
}
\arguments{
\item{theta}{Parameter of the Generalized Pareto function}

\item{c}{Parameter of the Generalized Pareto function}

\item{b}{Parameter of the Generalized Pareto function}

\item{rho}{Parameter of the Generalized Pareto function}

\item{delta}{Value of difference between iterations in order to output a result. Decreasing this parameter will increase accuracy of P0. Delta > 0}
}
\description{
Returns P0 of Generalized Pareto given a parameters theta, c, b , and rho
}
