% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Generalized_Pareto_calc_P0_iter}
\alias{Generalized_Pareto_calc_P0_iter}
\title{Generalized Pareto Distribution Function P0 with defined number of iterations}
\usage{
Generalized_Pareto_calc_P0_iter(theta, c, b, rho, iter)
}
\arguments{
\item{theta}{Parameter of the Generalized Pareto function}

\item{c}{Parameter of the Generalized Pareto function}

\item{b}{Parameter of the Generalized Pareto function}

\item{rho}{Parameter of the Generalized Pareto function}

\item{iter}{Number of iterations to be performed for summation calcuation. Increasing this parameter will increase accuracy of P0}
}
\description{
Returns P0 of Generalized Pareto given a parameters theta, c, b , and rho
}
