% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yule_generalized.R
\name{Generalized_Yule}
\alias{Generalized_Yule}
\title{Generalized Yule Distribution Function}
\usage{
Generalized_Yule(n, rho, alpha)
}
\arguments{
\item{n}{Length of vector to be generated.}

\item{rho}{Parameter of the Generalized Yule distribution function}

\item{alpha}{Parameter of the Generalized Yule distribution function: 0 <= alpha < 1}
}
\description{
This function generates a vector of n length of the Generalized Yule distribution with parameters rho and alpha.
}
\examples{
Generalized_Yule(100, 3, 0.1)
}
