% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{get_CI}
\alias{get_CI}
\title{Get Mean Confidence Interval Function}
\usage{
get_CI(data, alpha)
}
\arguments{
\item{data}{Data to get confidence interval from}

\item{alpha}{Alpha for confidence interval calculation}
}
\description{
This function generates a vector of confidence interval based on mean of data.
}
