% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{get_median_CI}
\alias{get_median_CI}
\title{Get Median Confidence Interval Function}
\usage{
get_median_CI(data)
}
\arguments{
\item{data}{Data to get confidence interval from}
}
\description{
This function generates a vector of ranked 95% confidence interval based on median of data.
}
