% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{get_p0}
\alias{get_p0}
\title{Psi Criterion for RGHD parameter ratios}
\usage{
get_p0(params, model_fn_name)
}
\arguments{
\item{params}{Vector of parameter for the model function}

\item{model_fn_name}{name of function as a character vector}
}
\description{
This function generates the Psi Criterion goodness of fit value given an empirical distribution for the 2m-RGHD function. Parameters r and q/r ratios are given, as well as desired weight of pmf and use of the weighted right-tail cumulative distribution function.
}
\examples{
params <- c(2, 3, 0.9)
get_p0(params, 'Kolmogorov Waring')
}
