% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{global_fit_RGHD_ratio}
\alias{global_fit_RGHD_ratio}
\title{Global optimization of the 2m-RGHD function given empirical data, r bounds, and q/r bounds.}
\usage{
global_fit_RGHD_ratio(
  param_bounds,
  data,
  iter,
  weighted_rt = F,
  n_cores = 1,
  clust
)
}
\arguments{
\item{param_bounds}{A list of sequences which indicate space where parameters should be generated and fit}

\item{data}{Vector of observed values}

\item{iter}{Integer indicating number of iterations to run grid search. Increasing iterations will increase decimal point precision of output parameters.}

\item{weighted_rt}{Boolean used to determine if the weighted right-tail cumulative distribution function should be used or not.}

\item{n_cores}{Integer used to indicate number of cores to be used for this function if a socket cluster object is not defined.}

\item{clust}{socket cluster object from 'parallel::makeCluster()'. This is used if you have already generated a socket cluster object and would like to run this functoin on it. If no object is defined, one will be made for this function call.}
}
\description{
This function generates a single set of optimized parameters and Psi Criterion for a given function within specified starting parameter bounds. This function uses Limited Memory BFGS as it's gradient descent algorithm.
}
\examples{
obs_data <- c(16362,7910,4563,2880,1992,1242,936,734,553,346,253,296,185,210,120,121,119,94,87,55,47,60,67,26,24,54,32,31,16,19,19,20,11,22,21,25,8,8,28,9,4,4,2,8,6,5,1,9,0,6,2,4,2,1,0,4,3,2,6,5,3,1,1,0,3,4,0,1,0,0,0,1,1,0,0,0,0,0,0,3,0,0,0,0,1,0,0,2,1,5,0,0,0,1,1,1,0,0,0,0,0,0,0,5,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4)
param_bounds <- list(seq(1,5,0.1),seq(1,5,0.1),seq(0.1,3,0.1),seq(0.1,3,0.1))
parameters <- global_fit_RGHD_ratio(param_bounds, obs_data, iter = 3, n_cores = 12)

#Or if a cluster is defined
cl <- makeCluster(12)
clusterCall(cl, function() library(tidyverse))
clusterCall(cl, function() library(optimr))
clusterCall(cl, function() library(SkeweDF))
parameters <- global_fit_RGHD_ratio(param_bounds, obs_data, iter = 3, clust = cl)
}
