% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Kolmogorov_Waring}
\alias{Kolmogorov_Waring}
\title{Kolmogorov Waring}
\usage{
Kolmogorov_Waring(n, a, b, theta)
}
\arguments{
\item{n}{Length of vector to be generated}

\item{a}{Vector of parameters of the Kolmogorov Waring distribution function}

\item{b}{Vector of parameters of the Kolmogorov Waring distribution function}

\item{theta}{Parameter of the Kolmogorov Waring distribution function}
}
\description{
Calculates vector of n length of Kolmogorov distribution function given parameters
}
