% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Kolmogorov_Waring_Moment}
\alias{Kolmogorov_Waring_Moment}
\title{Kolmogorov Waring Factorial Moment}
\usage{
Kolmogorov_Waring_Moment(a1, b1, theta, r)
}
\arguments{
\item{a1}{Parameter of the Kolmogorov Waring distribution function}

\item{b1}{Parameter of the Kolmogorov Waring distribution function}

\item{theta}{Parameter of the Kolmogorov Waring distribution function}

\item{r}{Index of factorial moment to calculate}
}
\description{
Calculates P0 of Kolmogorov Waring distribution function given parameters
}
