% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Kolmogorov_Waring_P0_calc}
\alias{Kolmogorov_Waring_P0_calc}
\title{Kolmogorov Waring P0 calculation}
\usage{
Kolmogorov_Waring_P0_calc(a, b, theta)
}
\arguments{
\item{a}{Parameter of the Kolmogorov Waring distribution function}

\item{b}{Parameter of the Kolmogorov Waring distribution function}

\item{theta}{Parameter of the Kolmogorov Waring distribution function}
}
\description{
Calculates P0 of Kolmogorov Waring distribution function given parameters
}
