% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{Lorentzian}
\alias{Lorentzian}
\title{Lorentzian Distribution Function}
\usage{
Lorentzian(n, gamma, x0, c)
}
\arguments{
\item{n}{Length of vector to be generated.}

\item{gamma}{Parameter of the Lorenzian distribution function}

\item{x0}{Parameter of the Lorenzian distribution function indicating center of function}

\item{c}{Parameter of the Lorenzian distribution function indicating center of function}
}
\description{
This function generates a vector of n length of the Lorentzian distribution
}
\examples{
Lorentzian_calc(5, 5.5, 6, 2)
}
