% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{Lorentzian_calc}
\alias{Lorentzian_calc}
\title{Lorentzian Distribution Function calculation}
\usage{
Lorentzian_calc(x, gamma, x0, c)
}
\arguments{
\item{x}{Index of function}

\item{gamma}{Parameter of the Lorenzian distribution function}

\item{x0}{Parameter of the Lorenzian distribution function indicating center of function}

\item{c}{Parameter of the Lorenzian distribution function indicating center of function}
}
\description{
This function calculates value of Lorentzian function at x
}
\examples{
Lorentzian_calc(5, 5.5, 6, 2)
}
