% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_to_mixed_model.R
\name{par_to_mixed_model}
\alias{par_to_mixed_model}
\title{Parameters to Mixed Model}
\usage{
par_to_mixed_model(n, x, d)
}
\arguments{
\item{n}{Int. Length of model vector}

\item{x}{Vector. Parameters}

\item{d}{Vector. Dimensions of mixed model}
}
\description{
Helper function which inputs a vector of parameters output from fit_mixed_kw and creates a model
}
