% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{parameter_post_processing}
\alias{parameter_post_processing}
\title{Parameter Optimization Helper Function}
\usage{
parameter_post_processing(parameter_df, model_fn_name, data)
}
\arguments{
\item{parameter_df}{Output dataframe of optimized parameters using local algorithm}

\item{model_fn_name}{Character vector used to indicate name of model function used for optimization}

\item{data}{Vector of observed values}
}
\description{
This function adds in additional columns to the optimized parameter output dataframe
}
