% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_criterion_generalized_kw.R
\name{psi_criterion_generalized_kw}
\alias{psi_criterion_generalized_kw}
\title{Psi Criterion Generalized KW}
\usage{
psi_criterion_generalized_kw(
  x,
  d,
  data,
  left_trunc,
  right_trunc,
  pmf_weight,
  weighted_rt,
  ratio = FALSE
)
}
\arguments{
\item{x}{Vector of parameters}

\item{d}{Int used to indicate the number of 'a' terms within the 'param_bounds' variable. The remaining values will be considered 'b' terms.}

\item{data}{Vector of observed values}

\item{left_trunc}{Int used to determine starting index of model to use for optimization}

\item{right_trunc}{Int used to determine ending index of model to use for optimization}

\item{pmf_weight}{Numeric of weight given to probability mass function for generation of Psi Criterion. For example, if pmf_weight <- 0.5, 50 percent of the Psi Criterion value will be attributed to the probability mass function while the other 50 percent will be attributed to the right-tail cumulative distribution function.}

\item{weighted_rt}{Boolean used to determine if the weighted right-tail cumulative distribution function should be used or not.}

\item{ratio}{Boolean used to determine if parameters provide are a/b ratios rather than absolute values.}
}
\description{
This function calculates the Psi criterion goodness of fit metric given a set of parameters for the Generalized Kolmogorov Waring function.
}
