% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_generalized_kw_cdf}
\alias{psi_criterion_generalized_kw_cdf}
\title{Psi Criterion Generalized KW CDF}
\usage{
psi_criterion_generalized_kw_cdf(
  x,
  m,
  data,
  left_trunc,
  right_trunc,
  weighted_rt
)
}
\description{
Psi Criterion Generalized KW CDF
}
