% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_generalized_kw_pmf}
\alias{psi_criterion_generalized_kw_pmf}
\title{Psi Criterion Generalized KW PMF}
\usage{
psi_criterion_generalized_kw_pmf(x, m, data, left_trunc, right_trunc)
}
\description{
Psi Criterion Generalized KW PMF
}
