% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_mixed_kw}
\alias{psi_criterion_mixed_kw}
\title{Psi Criterion Mixed KW}
\usage{
psi_criterion_mixed_kw(
  x,
  d,
  data,
  left_trunc,
  right_trunc,
  pmf_weight,
  weighted_rt
)
}
\description{
Psi Criterion Mixed KW
}
