% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_mixed_kw_pmf}
\alias{psi_criterion_mixed_kw_pmf}
\title{Psi Criterion Mixed KW PMF}
\usage{
psi_criterion_mixed_kw_pmf(x, d, data, left_trunc, right_trunc)
}
\description{
Psi Criterion Mixed KW PMF
}
