% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RGHD}
\alias{RGHD}
\title{2m-RGHD Distribution Function}
\usage{
RGHD(J, m, r, q, P0_iter = 100L, P0_included = FALSE)
}
\arguments{
\item{J}{Length of vector to be generated}

\item{m}{Parameter of the 2m-RGHD function, this defines number of r and q parameters of the function}

\item{r}{R vector containing r parameters from 1:m}

\item{q}{R vector containing q parameters from 1:m}

\item{P0_iter}{Integer indicating number of iterations to use for calculation of P0, increasing this parameter will increase accuracy of P0}

\item{P0_included}{Boolean used to include P0 in vector or not}
}
\description{
Returns doubly truncated vector of 2m-RGHD function values where input is 1-J
}
