% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RGHD_P0}
\alias{RGHD_P0}
\title{2m-RGHD Distribution Function P0}
\usage{
RGHD_P0(m, r, q)
}
\arguments{
\item{m}{Parameter of the 2m-RGHD function, this defines number of r and q parameters of the function}

\item{r}{R vector containing r parameters from 1:m}

\item{q}{R vector containing q parameters from 1:m}
}
\description{
Calculates P0 given a set of parameters
}
