% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_tail_cdf.R
\name{right_tail_cdf}
\alias{right_tail_cdf}
\title{Right-Tail Cumulative Distribution Function}
\usage{
right_tail_cdf(x)
}
\arguments{
\item{x}{Length of vector to be generated.}
}
\description{
This function generates a vector of the right-tail cumulative distribution function of a given vector of values.
}
\examples{
x <- c(1,2,3,4,5)
right_tail_cdf(x)
}
