% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{sample_to_data_table}
\alias{sample_to_data_table}
\title{Sample to Data Table}
\usage{
sample_to_data_table(sample)
}
\arguments{
\item{sample}{Vector}
}
\description{
Helper function which converts a sample vector to a frequency dataframe.
}
