% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{weighted_left_tail_cdf}
\alias{weighted_left_tail_cdf}
\title{Weighted Left-Tail Cumulative Distribution Function}
\usage{
weighted_left_tail_cdf(x)
}
\arguments{
\item{x}{Length of vector to be generated.}
}
\description{
This function generates a vector of the weighted left-tail cumulative distribution function of a given vector of values. The weight of of each variable is determined by its position in the vector. For example, with a vector of length 5, element 1 will have weight 5/(5+4+3+2+1). Element 1 will have weight 5/(5+4+3+2+1)
}
\examples{
x <- c(1,2,3,4,5)
weighted_left_tail_cdf(x)
}
