% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_tail_cdf_weighted.R
\name{weighted_right_tail_cdf}
\alias{weighted_right_tail_cdf}
\title{Weighted Right-Tail Cumulative Distribution Function}
\usage{
weighted_right_tail_cdf(x)
}
\arguments{
\item{x}{Length of vector to be generated.}
}
\description{
This function generates a vector of the weighted right-tail cumulative distribution function of a given vector of values. The weight of of each variable is determined by its position in the vector. For example, with a vector of length 5, element 5 will have weight 5/(5+4+3+2+1). Element 1 will have weight 1/(5+4+3+2+1)
}
\examples{
x <- c(1,2,3,4,5)
weighted_right_tail_cdf(x)
}
