% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_table.R
\name{write_input_table}
\alias{write_input_table}
\title{Write Input Table Helper Function}
\usage{
write_input_table(folder_name, data)
}
\arguments{
\item{folder_name}{Character vector indicating folder or directory name to be used when outputting table}

\item{data}{Vector of observed values}
}
\description{
This function generates table of input data
}
