% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parameter_table.R
\name{write_parameter_table}
\alias{write_parameter_table}
\title{Write Parameter Table Helper Function}
\usage{
write_parameter_table(parameter_df, folder_name, model_fn_name, RGHD_m = 0)
}
\arguments{
\item{parameter_df}{Data frame of optimized parameters and other model function values (p0, Psi, etc)}

\item{folder_name}{Character vector indicating folder or directory name to be used when outputting table}

\item{model_fn_name}{Character vector used to indicate name of model function used for optimization}

\item{RGHD_m}{Int indicating m value of 2m-RGHD function if applicable}
}
\description{
This function generates table of optimized parameters
}
