% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{Yule}
\alias{Yule}
\title{Yule Distribution Function}
\usage{
Yule(n, rho)
}
\arguments{
\item{n}{Length of vector to be generated.}

\item{rho}{Parameter of the Yule distribution function}
}
\description{
This function generates a vector of n length of the Yule distribution with parameter rho.
}
\examples{
Yule(100, 3)
}
