% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{append_predicted_values}
\alias{append_predicted_values}
\title{Append Predicted Values}
\usage{
append_predicted_values(
  data,
  model,
  left_trunc = 1,
  right_trunc = left_trunc + length(data) - 1
)
}
\arguments{
\item{data}{Vector of observed values}

\item{model}{Vector of optimized model}

\item{left_trunc}{Int used to determine starting index of model to use for optimization}

\item{right_trunc}{Int used to determine ending index of model to use for optimization}
}
\description{
This function will take in a left truncated data vector and append missing values according to the given model
}
