% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_generalized_kw}
\alias{psi_criterion_generalized_kw}
\title{Psi Criterion Generalized KW}
\usage{
psi_criterion_generalized_kw(
  x,
  m,
  data,
  left_trunc,
  right_trunc,
  pmf_weight,
  weighted_rt
)
}
\description{
Psi Criterion Generalized KW
}
