% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_generalized_kw_pmf}
\alias{psi_criterion_generalized_kw_pmf}
\title{Psi Criterion Generalized KW PMF}
\usage{
psi_criterion_generalized_kw_pmf(x, d, data, left_trunc, right_trunc)
}
\arguments{
\item{x}{Vector of parameters}

\item{d}{Int used to indicate the number of 'a' terms within the 'param_bounds' variable. The remaining values will be considered 'b' terms.}

\item{data}{Vector of observed values}

\item{left_trunc}{Int used to determine starting index of model to use for optimization}

\item{right_trunc}{Int used to determine ending index of model to use for optimization}
}
\description{
This function calculates the Psi criterion goodness of cit metric given a set of parameters for the probability mass function of the Generalized Kolmogorov Waring function.
}
