% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeweDF_functions.R
\name{psi_criterion_mixed_kw}
\alias{psi_criterion_mixed_kw}
\title{Psi Criterion Mixed KW}
\usage{
psi_criterion_mixed_kw(
  x,
  d,
  data,
  left_trunc,
  right_trunc,
  pmf_weight,
  weighted_rt
)
}
\arguments{
\item{x}{Vector of parameters}

\item{d}{Int used to indicate the number of 'a' terms within the 'param_bounds' variable. The remaining values will be considered 'b' terms.}

\item{data}{Vector of observed values}

\item{left_trunc}{Int used to determine starting index of model to use for optimization}

\item{right_trunc}{Int used to determine ending index of model to use for optimization}

\item{weighted_rt}{Boolean used to determine if the weighted right-tail cumulative distribution function should be used or not.}
}
\description{
This function calculates the Psi criterion goodness of fit metric given a set of parameters for a mixed KW function
}
